requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg[binary]==3.2.12
python-dotenv==1.0.1
gunicorn==22.0.0
whitenoise==6.7.0
django-filter==24.3
dj-database-url==2.2.0
pytest==8.3.2
pytest-django==4.9.0

